Stereo work.  Added interlaced back again.  Added interlaced columns.  Added checkerboard.
Former-commit-id: 8d5377e1e8c0b70ed46d24584fdc0a7198a835d9

+++ mrViewer/src/gui/mrvImageView.h
@@ -343,7 +343,7 @@ class server;
     void play_backwards();
 
     /// Scrub sequence
-    void scrub( float dy );
+    void scrub( double dy );
 
     /// Stop
     void stop();
